Refactored ContextMenu
