+++ src/components/canvas/Interaction/ContextMenu/index.ts
@@ -1,3 +1,6 @@
 import { ContextMenu } from './ContextMenu'
+import { useMenuCommands } from './useMenuCommands'
+import { MenuObjType, MenuInfo, MenuElement, MenuDescriptor } from './types'
 
-export { ContextMenu }
+export { ContextMenu, useMenuCommands }
+export type { MenuObjType, MenuInfo, MenuElement, MenuDescriptor }
