+++ src/components/canvas/Interaction/index.ts
@@ -1,6 +1,6 @@
 import { GeometryInteraction } from './GeometryInteraction'
 import { HighlightedObjects } from './HighlightedObjects'
-import { ContextMenu } from './ContextMenu'
+import { ContextMenu, useMenuCommands } from './ContextMenu'
 import { useOutlinesStore } from './OutlinesStore'
 
-export { GeometryInteraction, HighlightedObjects, ContextMenu, useOutlinesStore }
+export { GeometryInteraction, HighlightedObjects, ContextMenu, useMenuCommands, useOutlinesStore }
