+++ src/components/canvas/index.ts
@@ -5,6 +5,6 @@ import { Fit } from './Fit'
 import { Lights } from './Lights'
 import { raycastFilter } from './RaycastFilter'
 import { Threshold } from './Threshold'
-import { GeometryInteraction, HighlightedObjects, ContextMenu } from './Interaction'
+import { GeometryInteraction, HighlightedObjects, ContextMenu, useMenuCommands } from './Interaction'
 
-export { AutoClear, Fit, Lights, Controls, Composer, raycastFilter, Threshold, GeometryInteraction, HighlightedObjects, ContextMenu }
+export { AutoClear, Fit, Lights, Controls, Composer, raycastFilter, Threshold, GeometryInteraction, HighlightedObjects, ContextMenu, useMenuCommands }
